+++ sdk/hdinsight/arm-hdinsight/src/models/virtualMachinesMappers.ts
@@ -8,6 +8,5 @@
 
 export {
   ErrorResponse,
-  HostInfo,
-  HostInfoListResult
+  HostInfo
 } from "../models/mappers";