+++ sdk/identity/identity/src/credentials/chainedTokenCredential.ts
@@ -11,6 +11,11 @@ import { CanonicalCode } from "@opentelemetry/api";
  * until one of the getToken methods returns an access token.
  */
 export class ChainedTokenCredential implements TokenCredential {
+  /**
+   * The message to use when the chained token fails to get a token
+   */
+  protected UnavailableMessage =
+    "ChainedTokenCredential failed to retrieve a token from the included credentials";
   private _sources: TokenCredential[] = [];
 
   /**