+++ sdk/eventhub/event-hubs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 5.2.2 (Unreleased)
+
+- Fixes issue [#9289](https://github.com/Azure/azure-sdk-for-js/issues/9289)
+  where calling `await subscription.close()` inside of a subscription's `processError`
+  handler would cause the subscription to deadlock.
+
 ## 5.2.1 (2020-06-08)
 
 - Fixes issue [#8584](https://github.com/Azure/azure-sdk-for-js/issues/8584)